+++ CHANGELOG.rst
@@ -9,6 +9,8 @@ By import convention, components of the Sciris library are listed beginning with
 Version 1.1.2 (2021-07-05)
 --------------------------
 1. Updated ``sc.heading()`` to handle arguments the same way as ``print()``, e.g. ``sc.heading([1,2,3], 'is a list')``.
+1. Updated ``sc.commaticks()`` to use locale-aware formatting. 
+1. Removed the ``fig`` argument from ``sc.commaticks()`` and ``sc.SIticks()``; now, the first argument can be an ``Axes`` object, a ``Figure`` object, or a list of axes.
 
 
 Version 1.1.1 (2021-03-17)